fix lib/firebase
